Pass merge-fix distance as a configuration option
So I can play around with it. On the one hand, pass something low and
there are fewer to check manually. On the other hand, if you are dealing
with more subtle bugs, then perhaps they would not show up (and be
fixed) till much later either...

+++ src/find_bug_fix.rs
@@ -157,7 +157,7 @@ pub fn within_n_generations(
     repo: &git2::Repository,
     root: &git2::Oid,
     child: &git2::Oid,
-    n: usize,
+    n: u32,
 ) -> bool {
     let child = repo.find_commit(*child).unwrap();
     let mut children = vec![child];
